# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"llmdex","summary":"Turns a Hugging Face model ID into a reviewed specification sheet.","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Health"}}}}}}},"/models/{model_id}/agents/stream":{"get":{"tags":["agents"],"summary":"Stream Agents","description":"Run the named agents and narrate them as server-sent events.\n\nEverything that can be refused is refused before the stream opens: once the\nresponse has begun, a 404 can no longer be sent, and an error buried in the\nbody is one a client has to know to look for.","operationId":"stream_agents_models__model_id__agents_stream_get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}},{"name":"agents","in":"query","required":false,"schema":{"type":"string","description":"comma-separated agent names","default":"about,prose,benchmarks","title":"Agents"},"description":"comma-separated agent names"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ingest":{"post":{"tags":["ingest"],"summary":"Ingest Model","description":"Fetch, derive, and write a document. Atomic: it completes or it fails (R1.5).\n\nA model entering the vault for the first time is summarised on the way in, so\nnobody has to ask for the first one. That step cannot fail this endpoint: see\n:func:`~app.features.summary.router.summarise_after_first_ingest`.","operationId":"ingest_model_ingest_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDoc"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/models":{"get":{"tags":["models"],"summary":"List Models","description":"R6.6 - has the upstream card moved since we read it?","operationId":"list_models_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ModelDoc"},"type":"array","title":"Response List Models Models Get"}}}}}}},"/models/{model_id}/drift":{"get":{"tags":["models"],"summary":"Model Drift","description":"R6.6 - has the upstream card moved since we read it?","operationId":"model_drift_models__model_id__drift_get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DriftReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/models/{model_id}":{"get":{"tags":["models"],"summary":"Get Model","description":"Every field, including the null ones (R6.3).","operationId":"get_model_models__model_id__get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDoc"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["models"],"summary":"Delete Model","description":"Remove a model from the vault.\n\n204 rather than the deleted document: there is nothing left to return, and a\nbody would invite a caller to treat it as still being there. The removal is\na commit in the vault repository, so this is undoable outside the app (R4.7).","operationId":"delete_model_models__model_id__delete","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/benchmarks":{"get":{"tags":["benchmarks"],"summary":"List Benchmarks","operationId":"list_benchmarks_benchmarks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Benchmark"},"type":"array","title":"Response List Benchmarks Benchmarks Get"}}}}}}},"/benchmarks/{slug}":{"get":{"tags":["benchmarks"],"summary":"Get Benchmark","operationId":"get_benchmark_benchmarks__slug__get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Benchmark"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/models/{model_id}/extract":{"post":{"tags":["extraction"],"summary":"Extract Model","description":"Read the card, ask the model to locate facts in it, store what verifies.\n\nA run that verifies nothing is a success: most cards state none of this, and\nan empty result with its rejections is information (R3.2).","operationId":"extract_model_models__model_id__extract_post","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDoc"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/models/{model_id}/summarize":{"post":{"tags":["summary"],"summary":"Summarise Model","description":"Read the card, search the web, write an account of the model.\n\nUnlike extraction this replaces what was there: regenerating is the point of\nthe button, and ``generated_on`` records which run produced the text on\nscreen.","operationId":"summarise_model_models__model_id__summarize_post","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDoc"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Benchmark":{"properties":{"slug":{"type":"string","title":"Slug"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"group":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Group"},"unit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unit"},"direction":{"anyOf":[{"type":"string","enum":["higher_is_better","lower_is_better"]},{"type":"null"}],"title":"Direction"},"explanation":{"type":"string","title":"Explanation","default":""},"unwritten":{"type":"boolean","title":"Unwritten","default":true},"referring_models":{"items":{"type":"string"},"type":"array","title":"Referring Models"}},"type":"object","required":["slug"],"title":"Benchmark","description":"R5.1 - one document per benchmark."},"BenchmarkScore":{"properties":{"slug":{"type":"string","title":"Slug"},"score":{"type":"number","title":"Score"},"unit":{"type":"string","title":"Unit"},"provenance":{"type":"string","enum":["vendor","internal","third_party"],"title":"Provenance"},"source_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Url"},"harness_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Harness Name"},"num_shots":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Num Shots"},"eval_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Eval Date"}},"type":"object","required":["slug","score","unit","provenance"],"title":"BenchmarkScore"},"Checkpoint":{"properties":{"repo":{"type":"string","title":"Repo"},"quantization":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Quantization"},"card_revision":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Card Revision"},"ingested":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ingested"},"derived":{"anyOf":[{"$ref":"#/components/schemas/Derived"},{"type":"null"}]},"extracted":{"anyOf":[{"$ref":"#/components/schemas/Extracted"},{"type":"null"}]},"extracted_benchmarks":{"anyOf":[{"$ref":"#/components/schemas/ExtractedBenchmarks"},{"type":"null"}]},"manual":{"$ref":"#/components/schemas/Manual"},"benchmarks":{"items":{"$ref":"#/components/schemas/BenchmarkScore"},"type":"array","title":"Benchmarks"},"measured":{"items":{"$ref":"#/components/schemas/Measured"},"type":"array","title":"Measured"}},"type":"object","required":["repo"],"title":"Checkpoint","description":"R4.1 - one artifact of a model: a specific quantization, in a specific repo."},"Derived":{"properties":{"architecture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Architecture"},"architecture_class":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Architecture Class"},"hidden_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Hidden Size"},"num_hidden_layers":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Num Hidden Layers"},"num_attention_heads":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Num Attention Heads"},"num_key_value_heads":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Num Key Value Heads"},"head_dim":{"$ref":"#/components/schemas/HeadDim"},"vocab_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Vocab Size"},"context_length":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Context Length"},"torch_dtype":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Torch Dtype"},"layers":{"$ref":"#/components/schemas/LayerComposition"},"params":{"$ref":"#/components/schemas/ParamCounts"},"weights":{"$ref":"#/components/schemas/WeightBytes"},"vram":{"anyOf":[{"$ref":"#/components/schemas/VRAMEstimate"},{"type":"null"}]},"underivable":{"items":{"type":"string"},"type":"array","title":"Underivable"}},"type":"object","title":"Derived","description":"R2.1 - fields computed from structured files. Never a guess."},"DriftReport":{"properties":{"model_id":{"type":"string","title":"Model Id"},"stored_revision":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stored Revision"},"upstream_revision":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Upstream Revision"},"drifted":{"type":"boolean","title":"Drifted"}},"type":"object","required":["model_id","stored_revision","upstream_revision","drifted"],"title":"DriftReport"},"Extracted":{"properties":{"card_revision":{"type":"string","title":"Card Revision"},"extracted_on":{"type":"string","title":"Extracted On"},"model":{"type":"string","title":"Model"},"quantization":{"anyOf":[{"$ref":"#/components/schemas/ExtractedQuantization"},{"type":"null"}]},"serving":{"anyOf":[{"$ref":"#/components/schemas/ExtractedServing"},{"type":"null"}]},"rejected":{"items":{"$ref":"#/components/schemas/RejectedValue"},"type":"array","title":"Rejected"}},"additionalProperties":false,"type":"object","required":["card_revision","extracted_on","model"],"title":"Extracted","description":"R3.x - what a language model found in the card, after verification.\n\n``card_revision`` is carried here rather than borrowed from ingest because a\nspan is only meaningful against one specific text. If the card moves, these\noffsets describe the revision this block names, not the current one."},"ExtractedBenchmark":{"properties":{"name":{"$ref":"#/components/schemas/Span"},"score":{"$ref":"#/components/schemas/Span"},"unit":{"anyOf":[{"$ref":"#/components/schemas/Span"},{"type":"null"}]},"variant":{"anyOf":[{"$ref":"#/components/schemas/Span"},{"type":"null"}]}},"additionalProperties":false,"type":"object","required":["name","score"],"title":"ExtractedBenchmark","description":"R3.4 - one cell of the vendor's reported table.\n\n``score`` stays a string. Turning ``\"52.80\"`` into ``52.8`` is a conversion,\nand R3.1 forbids conversion at extraction time; that promotion happens with\nthe Benchmarks tab, where a human confirms the slug (R4.5b, R5.3).\n\n``variant`` is the column header the score sat under, copied like everything\nelse. Published tables compare checkpoints -- a BF16 sibling beside this\nNVFP4 one -- and without the header a number from the wrong column reads as\nthis repository's result. Nothing here decides which column *is* this\ncheckpoint: the header text and a repo name are close but not equal, and\nthat gap is exactly where a wrong attribution would live."},"ExtractedBenchmarks":{"properties":{"card_revision":{"type":"string","title":"Card Revision"},"extracted_on":{"type":"string","title":"Extracted On"},"model":{"type":"string","title":"Model"},"rows":{"items":{"$ref":"#/components/schemas/ExtractedBenchmark"},"type":"array","title":"Rows"},"rejected":{"items":{"$ref":"#/components/schemas/RejectedValue"},"type":"array","title":"Rejected"}},"additionalProperties":false,"type":"object","required":["card_revision","extracted_on","model"],"title":"ExtractedBenchmarks","description":"R3.4 - the results table of one card, as its own block.\n\nSeparate from :class:`~app.features.extraction.schemas.Extracted` because a\nseparate agent reads it. Both blocks carry the revision of the card their\nown spans were found in, and a re-run of one cannot overwrite the other."},"ExtractedQuantization":{"properties":{"format":{"anyOf":[{"$ref":"#/components/schemas/Span"},{"type":"null"}]},"method":{"anyOf":[{"$ref":"#/components/schemas/Span"},{"type":"null"}]},"scope":{"anyOf":[{"$ref":"#/components/schemas/Span"},{"type":"null"}]},"calibration":{"anyOf":[{"$ref":"#/components/schemas/Span"},{"type":"null"}]}},"additionalProperties":false,"type":"object","title":"ExtractedQuantization","description":"R3.4 - the same four fields as ``Quantization``, each as a verified span."},"ExtractedServing":{"properties":{"engines":{"additionalProperties":{"$ref":"#/components/schemas/Span"},"type":"object","title":"Engines"}},"additionalProperties":false,"type":"object","title":"ExtractedServing"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HeadDim":{"properties":{"value":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Value"},"source":{"anyOf":[{"type":"string","enum":["explicit","derived"]},{"type":"null"}],"title":"Source"},"mismatch":{"anyOf":[{"$ref":"#/components/schemas/HeadDimMismatch"},{"type":"null"}]}},"type":"object","title":"HeadDim"},"HeadDimMismatch":{"properties":{"explicit":{"type":"integer","title":"Explicit"},"derived":{"type":"integer","title":"Derived"}},"type":"object","required":["explicit","derived"],"title":"HeadDimMismatch","description":"R2.4a - config states one head_dim, hidden/heads implies another."},"Health":{"properties":{"status":{"type":"string","title":"Status"},"vault":{"type":"string","title":"Vault"},"vault_exists":{"type":"boolean","title":"Vault Exists"}},"type":"object","required":["status","vault","vault_exists"],"title":"Health"},"IngestRequest":{"properties":{"model_id":{"type":"string","title":"Model Id","description":"A Hugging Face model ID or a full URL, e.g. `Qwen/Qwen3-8B`.","examples":["Qwen/Qwen3-8B"]},"context":{"type":"integer","exclusiveMinimum":0.0,"title":"Context","description":"Context length the VRAM estimate is computed at (R2.5).","default":32768}},"type":"object","required":["model_id"],"title":"IngestRequest"},"LayerComposition":{"properties":{"family":{"type":"string","enum":["transformer","nemotron_h","jamba","hybrid","recurrent","unknown"],"title":"Family","default":"transformer"},"attention":{"type":"integer","title":"Attention","default":0},"recurrent":{"type":"integer","title":"Recurrent","default":0},"mlp_only":{"type":"integer","title":"Mlp Only","default":0},"recurrent_kind":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recurrent Kind"},"unreliable_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unreliable Reason"}},"type":"object","title":"LayerComposition","description":"How many layers of each kind. R2.6a - never assume a fixed ratio.\n\n``family`` names the convention the config used to declare its layout, not\nthe vendor. ``nemotron_h`` and ``jamba`` are kept as their own values because\ndocuments written before the others existed carry them."},"Manual":{"properties":{"reviewed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reviewed"},"quantization":{"anyOf":[{"$ref":"#/components/schemas/Quantization"},{"type":"null"}]},"serving":{"anyOf":[{"$ref":"#/components/schemas/Serving"},{"type":"null"}]}},"additionalProperties":false,"type":"object","title":"Manual","description":"R6.5 - hand-entered prose. Ingest never writes here.\n\n``reviewed`` separates two of R6.4's four states. Null means nobody has read\nthe card, so a null field below it is unknown. Set means a person read it, so\na null field below it is genuinely absent from the card.\n\n``extra=\"forbid\"`` matters because this block is edited by hand: Pydantic's\ndefault would drop a mistyped key silently and the next write would erase it."},"Measured":{"properties":{"hardware":{"type":"string","minLength":1,"title":"Hardware"},"serving":{"type":"string","minLength":1,"title":"Serving"},"ttft_ms":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Ttft Ms"},"itl_ms":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Itl Ms"},"throughput_tok_s":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Throughput Tok S"},"peak_vram_bytes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Peak Vram Bytes"},"measured_on":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Measured On"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["hardware","serving"],"title":"Measured","description":"R4.2 / R4.3 - numbers the team produces. Ingest never writes these.\n\nHardware and serving configuration are required, not optional context. A\nthroughput figure without them describes nothing anyone can act on, so a\npartial entry is rejected rather than stored."},"ModelDoc":{"properties":{"model_id":{"type":"string","title":"Model Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"vendor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vendor"},"released":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Released"},"summary":{"anyOf":[{"$ref":"#/components/schemas/Summary"},{"type":"null"}]},"checkpoints":{"items":{"$ref":"#/components/schemas/Checkpoint"},"type":"array","title":"Checkpoints"}},"type":"object","required":["model_id"],"title":"ModelDoc","description":"R4.1 - shared identity; everything that varies by artifact lives on checkpoints."},"ParamCounts":{"properties":{"total":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total"},"active":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Active"},"is_moe":{"type":"boolean","title":"Is Moe","default":false},"auxiliary":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Auxiliary"},"auxiliary_module":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Auxiliary Module"},"unreliable_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unreliable Reason"}},"type":"object","title":"ParamCounts","description":"R2.2 - how many weights the model has, and how many fire per token.\n\nCarries ``unreliable_reason`` for the same reason every other block here\ndoes: both numbers can be wrong for structural reasons rather than missing\nones -- a quantized checkpoint whose reported total counts packed bytes, or\na hybrid whose expert layers are a subset of its layers -- and a bare null\nrecords that we have no number without recording that we knew why."},"Quantization":{"properties":{"format":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Format"},"method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Method"},"scope":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scope"},"calibration":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Calibration"},"src":{"type":"string","minLength":1,"title":"Src"}},"additionalProperties":false,"type":"object","required":["src"],"title":"Quantization","description":"R3.4 - the quantization recipe, copied from the card by hand.\n\n``src`` is required (R3.3). A recipe with no pointer at the section it came\nfrom cannot be re-checked when the card changes underneath it, and an\nunauditable value is the thing this project exists to not produce."},"RejectedValue":{"properties":{"field":{"type":"string","title":"Field"},"proposed":{"type":"string","title":"Proposed"},"reason":{"type":"string","enum":["no_match","empty","not_contiguous"],"title":"Reason"}},"additionalProperties":false,"type":"object","required":["field","proposed","reason"],"title":"RejectedValue","description":"R3.2 - what the model claimed, and why it was not stored.\n\nA bare null would lose the fact that the model produced something. This is\nwhere invention becomes visible, and it is the main evidence for whether an\nendpoint can be trusted with this job at all."},"Serving":{"properties":{"engines":{"additionalProperties":{"type":"string"},"type":"object","minProperties":1,"title":"Engines"},"src":{"type":"string","minLength":1,"title":"Src"}},"additionalProperties":false,"type":"object","required":["engines","src"],"title":"Serving","description":"R3.4 - per-engine support with version pins, exactly as the card states them.\n\nValues stay strings because a card pins a version (``\"0.27.1\"``) or states a\ncondition (``\"dev container only\"``). Both are facts worth keeping; parsing\nwould force the second into a shape it does not have."},"Span":{"properties":{"text":{"type":"string","title":"Text"},"start":{"type":"integer","title":"Start"},"end":{"type":"integer","title":"End"},"section":{"type":"string","title":"Section"},"occurrences":{"type":"integer","title":"Occurrences","default":1}},"additionalProperties":false,"type":"object","required":["text","start","end","section"],"title":"Span","description":"A verified quote and where it was found in the card.\n\nConstructed only after :mod:`app.core.grounding` has matched the quote\nagainst the source, which is why every field is required. ``text`` is sliced\nout of the card rather than copied from the model's response -- the model\nlocates, it does not supply (R3.1)."},"Summary":{"properties":{"overview":{"type":"string","minLength":1,"title":"Overview"},"unique_points":{"items":{"type":"string"},"type":"array","title":"Unique Points"},"pros":{"items":{"type":"string"},"type":"array","title":"Pros"},"cons":{"items":{"type":"string"},"type":"array","title":"Cons"},"use_cases":{"items":{"type":"string"},"type":"array","title":"Use Cases"},"sources":{"items":{"type":"string"},"type":"array","title":"Sources"},"generated_by":{"type":"string","title":"Generated By"},"generated_on":{"type":"string","title":"Generated On"}},"additionalProperties":false,"type":"object","required":["overview","generated_by","generated_on"],"title":"Summary","description":"A written account of what the model is, generated rather than copied.\n\nThis is the one block in the system that is not quoted from somewhere. It\nexists because the copy-only rule (R3.1), correct for a specification sheet,\ncannot answer \"what is this model good for\" -- no card states its own cons.\nIt is therefore held to a different discipline: not verbatim, but sourced,\ndated, attributed to the endpoint that wrote it, and regenerable.\n\nIt lives on the model rather than the checkpoint because the card and the\nfacts it draws on do not vary by quantization: an AWQ build and a GGUF build\nof one model are the same model to a reader asking what it is for."},"VRAMAssumptions":{"properties":{"context":{"type":"integer","title":"Context"},"batch":{"type":"integer","title":"Batch","default":1},"kv_dtype":{"type":"string","enum":["fp32","fp16","bf16","fp8","int8"],"title":"Kv Dtype","default":"fp16"},"overhead_bytes":{"type":"integer","title":"Overhead Bytes","default":2147483648}},"type":"object","required":["context"],"title":"VRAMAssumptions","description":"R2.5 - an estimate without these is not acceptable output."},"VRAMEstimate":{"properties":{"total_bytes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Bytes"},"weight_bytes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Weight Bytes"},"kv_cache_bytes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Kv Cache Bytes"},"assumptions":{"$ref":"#/components/schemas/VRAMAssumptions"},"unreliable_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unreliable Reason"}},"type":"object","required":["assumptions"],"title":"VRAMEstimate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WeightBytes":{"properties":{"bytes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Bytes"},"file_count":{"type":"integer","title":"File Count","default":0},"estimated":{"type":"boolean","title":"Estimated","default":false},"unreliable_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unreliable Reason"}},"type":"object","title":"WeightBytes"}}}}